mengganti UI

+++ database/migrations/2026_05_23_102900_add_username_to_users_table.php
@@ -8,19 +8,29 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'username')) {
 
-            $table->string('username')->unique();
+            Schema::table('users', function (Blueprint $table) {
 
-        });
+                $table->string('username')
+                    ->nullable()
+                    ->unique();
+
+            });
+
+        }
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'username')) {
 
-            $table->dropColumn('username');
+            Schema::table('users', function (Blueprint $table) {
 
-        });
+                $table->dropColumn('username');
+
+            });
+
+        }
     }
 };
